split path_info on NotADirectory
